feat(oidc): add missing overridable parameters from the request object attribute
closes gravitee-io/issues#4033
